% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{add_branch_protect}
\alias{add_branch_protect}
\title{Add branch protections}
\usage{
add_branch_protect(prefixed_name, branch = "main")
}
\arguments{
\item{prefixed_name}{Name of the project including the prefix}

\item{branch}{Name of the github branch to protect}
}
\description{
Add branch protections
}
\keyword{internal}
