% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{add_github_actions_offline}
\alias{add_github_actions_offline}
\title{Offline version of adding in github actions}
\usage{
add_github_actions_offline()
}
\description{
Offline version of adding in github actions
}
\keyword{internal}
