% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbankr_functions.R
\name{call_perl}
\alias{call_perl}
\title{Call system perl file
Create and run call for system perl file}
\usage{
call_perl(laste_dbase, laste_bruker, lastefilsti, lastefil, passord)
}
\arguments{
\item{laste_dbase}{Database for loading up file. Choose between "TEST", "QA" or "PROD"}

\item{laste_bruker}{Oracle use ID for loading to StatBank}

\item{lastefilsti}{Path to where the .dat file is located}

\item{lastefil}{Name of the .dat file for loading (case sensitive)}

\item{passord}{String variable with the password for the laste_bruker}
}
\description{
Call system perl file
Create and run call for system perl file
}
\keyword{internal}
