% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbankr_functions.R
\name{call_shell}
\alias{call_shell}
\title{System shell call
Create and run call to shell file}
\usage{
call_shell(
  laste_dbase,
  laste_bruker,
  hovedtabell,
  publiseringsdato,
  mailto,
  passord,
  autooverskriv = 1,
  autogodkjenn = 1
)
}
\arguments{
\item{laste_dbase}{Database for loading up file. Choose between "TEST", "QA" or "PROD"}

\item{laste_bruker}{Oracle use ID for loading to StatBank}

\item{hovedtabell}{Name of the tabell for publication}

\item{publiseringsdato}{Dato for publisering}

\item{mailto}{Initialer av bruker for å sende epost til}

\item{autooverskriv}{If dubbletter should be overwritten. 0 = no, 1 = yes (default)}

\item{autogodkjenn}{How data should be approved. 0 = Manual approval, 1 = Automatic (default), 2 = Just-in-time}
}
\description{
System shell call
Create and run call to shell file
}
\keyword{internal}
