% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testtemplate.R
\name{create_my_project}
\alias{create_my_project}
\title{Create a new project from my minimal template}
\usage{
create_my_project(path)
}
\arguments{
\item{path}{The path to create the new project}
}
\description{
Create a new project from my minimal template
}
