% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{create_project_file}
\alias{create_project_file}
\title{Create project file}
\usage{
create_project_file(path, prefixed_name, project_type = "package")
}
\arguments{
\item{path}{to where the new project is being set up}

\item{prefixed_name}{Name of the package including the ssb- or stat- prefix}

\item{project_type}{Whether the project is a package ("package") or project ("project").}
}
\description{
Create project file
}
\keyword{internal}
