% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynarev_uttrekk.R
\name{dynarev_uttrekk}
\alias{dynarev_uttrekk}
\title{Uttrekk fra Dynarev til R}
\usage{
dynarev_uttrekk(
  delregnr = getPass:::readline_nomask("Skriv inn delregnr: ", silent = TRUE),
  skjema = getPass:::readline_nomask("Skriv inn skjemanavn: ", silent = TRUE),
  enhets_type = getPass:::readline_nomask("Skriv inn enhets_type (BEDR/FRTK): ", silent
    = TRUE),
  skjema_cols = T,
  skjema_sfu_merge = F,
  dublettsjekk = F,
  sfu_cols = F,
  con_ask = T
)
}
\arguments{
\item{delregnr}{Numerisk vektor med delregisternummer}

\item{skjema}{Karakter vektor med skjemanavn}

\item{enhets_type}{Karakter vektor med enhetstype}

\item{skjema_cols}{TRUE = alle kolonner fra skjema. Karakter vektor for kun utvalgte kolonner}

\item{skjema_sfu_merge}{TRUE = skjemadata og SFU merges sammen til én data frame}

\item{dublettsjekk}{TRUE = dublettsjekk etter ENHETS_ID. Karakter vektor for dublettsjekk etter valgte kolonner.}

\item{sfu_cols}{TRUE = alle kolonner fra SFU. Karakter vektor for kun utvalgte kolonner}

\item{con_ask}{TRUE = spør om passord for å koble til DB1P og laster inn data. Dersom con_ask = "con" kobles det til DB1P uten å laste inn data}

\item{raadata}{FALSE returnerer reviderte data, TRUE returnerer rådata}
}
\value{

}
\description{
Uttrekk fra Dynarev til R
}
\examples{
# Det er mulig å bruke funksjonen uten å skrive inn noen argumenter. Man blir da bedt om å skrive inn skjemanavn, delregnr og enhets_type. Alle kolonnene fra skjemadata blir med.
dynarev <- dynarev_uttrekk()

# Her får man lastet inn alle kolonnene fra valgt skjema.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T,
                          enhets_type = "FRTK")

# Her får man lastet ned utvalgte kolonner fra valgt skjema. Her er AARGANG, FORETAKETS_NAVN og FORETAKETS_ORGNR valgt.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = c("AARGANG", "FORETAKETS_NAVN", "FORETAKETS_ORGNR"),
                          enhets_type = "FRTK")

# Her får man lastet ned alle kolonnene med SFU-data fra valgt skjema.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = F,
                          enhets_type = "FRTK",
                          sfu_cols = T)

# Her får man lastet ned valgte kolonner med SFU-data fra valgt skjema. Her er NAVN og ORGNR valgt.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = F,
                          enhets_type = "FRTK",
                          dublettsjekk = F,
                          sfu_cols = c("NAVN", "ORGNR"))

# Her får man lastet ned alle kolonnene med SFU-data fra alle skjema i valgt delregister. Dersom man kun ønsker utvalgte variabler fra SFU-data erstattes sfu_cols = T med sfu_cols = c("NAVN", "ORGNR").
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = T,
                          skjema_cols = F,
                          enhets_type = "FRTK",
                          sfu_cols = T)

# Her får man lastet ned både skjemadata (alle kolonner) og SFU-data (alle kolonner) fra valgt delregister. Objektet dynarev inneholder her altså to datasett. For å få hentet ut skjemadata skriver man dynarev_skjema <- data.frame(dynarev[1]) og for å hente ut SFU-data skriver man dynarev_sfu <- data.frame(dynarev[2]).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T,
                          enhets_type = "FRTK",
                          sfu_cols = T)
dynarev_skjema <- data.frame(dynarev[1])
dynarev_sfu <- data.frame(dynarev[2])

# Her får man lastet ned både skjemadata (alle kolonner) og SFU-data (utvalgte kolonner, her NAVN og ORGNR) fra valgt delregister. Objektet dynarev inneholder her altså to datasett. For å få hentet ut skjemadata skriver man dynarev_skjema <- data.frame(dynarev[1]) og for å hente ut SFU-data skriver man dynarev_sfu <- data.frame(dynarev[2]).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T,
                          enhets_type = "FRTK",
                          sfu_cols = c("NAVN", "ORGNR"))
dynarev_skjema <- data.frame(dynarev[1])
dynarev_sfu <- data.frame(dynarev[2])

# Her får man lastet ned både skjemadata (utvalgte kolonner, her AARGANG, FORETAKETS_NAVN og FORETAKETS_ORGNR) og SFU-data (utvalgte kolonner, her NAVN og ORGNR) fra valgt delregister. Objektet dynarev inneholder her altså to datasett. For å få hentet ut skjemadata skriver man dynarev_skjema <- data.frame(dynarev[1]) og for å hente ut SFU-data skriver man dynarev_sfu <- data.frame(dynarev[2]).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = c("AARGANG", "FORETAKETS_NAVN", "FORETAKETS_ORGNR"),
                          enhets_type = "FRTK",
                          sfu_cols = c("NAVN", "ORGNR"))
dynarev_skjema <- data.frame(dynarev[1])
dynarev_sfu <- data.frame(dynarev[2])

# Her får man lastet ned både skjemadata (alle kolonner) og SFU-data (alle kolonner) fra valgt delregister. SFU-dataene blir koblet på skjemadataene slik at man kun får én fil.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T,
                          skjema_sfu_merge = T,
                          enhets_type = "FRTK",
                          sfu_cols = T)

# Her får man lastet ned både skjemadata (alle kolonner) og SFU-data (utvalgte kolonner, her NAVN og ORGNR) fra valgt delregister. SFU-dataene blir koblet på skjemadataene slik at man kun får én fil.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T,
                          skjema_sfu_merge = T,
                          enhets_type = "FRTK",
                          sfu_cols = c("NAVN", "ORGNR"))

# Her får man lastet ned både skjemadata (utvalgte kolonner, her AARGANG, FORETAKETS_NAVN og FORETAKETS_ORGNR) og SFU-data (utvalgte kolonner, her NAVN og ORGNR) fra valgt delregister. SFU-dataene blir koblet på skjemadataene slik at man kun får én fil.
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = c("AARGANG", "FORETAKETS_NAVN", "FORETAKETS_ORGNR"),
                          skjema_sfu_merge = T,
                          enhets_type = "FRTK",
                          sfu_cols = c("NAVN", "ORGNR"))

# Her får man testet for dubletter (etter ENHETS_ID) i skjemadata. Liste med to datasett returneres; [1] skjemadata og [2] dublettdata (dersom det finnes dubletter, hvis ikke er denne blank).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                           skjema = "HELSE39",
                           enhets_type = "BEDR",
                           dublettsjekk = T)

# Dersom man ønsker å sjekke for dubletter etter én eller flere selvvalgte variabler skrives disse i en vektor, f.eks. c("variabel1", "variabel2").
dynarev <- dynarev_uttrekk(delregnr = 2421,
                           skjema = "HELSE0X",
                           enhets_type = "FRTK",
                           dublettsjekk = c("FORETAKSNR", "ART_SEKTOR", "FUNKSJON_KAPITTEL"))

# For å slippe å oppgi passord flere ganger når man skal laste inn flere skjema kan man skrive con_ask = FALSE. For at dette skal fungere må man logge på først øverst i scriptet: con <- dynarev_uttrekk(con_ask = "con")
con <- dynarev_uttrekk(con_ask = "con")

dynarev_1 <- dynarev_uttrekk(delregnr = 2421,
                             skjema = "HELSE41",
                             skjema_cols = T,
                             enhets_type = "FRTK",
                             con_ask = FALSE)

dynarev_2 <- dynarev_uttrekk(delregnr = 2421,
                             skjema = "HELSE46",
                             skjema_cols = T,
                             enhets_type = "BEDR",
                             con_ask = FALSE)

# For å laste inn alle skjema fra samme delregister i samme fil skriver man skjema = T. Under enhets_type må man samtidig skrive inn alle enhetstypene som finnes i det valgte delregisteret (f.eks. BEDR og FRTK)
dynarev <- dynarev_uttrekk(delregnr = 2420,
                          skjema = T,
                          skjema_cols = T,
                          enhets_type = c("BEDR", "FRTK"),
                          sfu_cols = F)

# Her lastes det inn SFU-data fra flere skjema (fra samme delregister).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = c("HELSE38P", "HELSE39", "HELSE44P"),
                          skjema_cols = F,
                          enhets_type = c("BEDR", "FRTK"),
                          sfu_cols = T)

# Her lastes det inn SFU-data fra alle skjema (fra samme delregister).
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = T,
                          skjema_cols = F,
                          enhets_type = c("BEDR", "FRTK"),
                          sfu_cols = c("NAVN", "ORGNR"))
                          
# Her lastes det inn rådata fra de aktive enhetene i valgt skjema.
dynarev <- dynarev_uttrekk(delregnr = delregnr, 
                           skjema = "HELSE0X",
                           skjema_cols = T,
                           enhets_type = c("BEDR", "FRTK"), 
                           raadata = T)
}
