% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{esquisser_ssb}
\alias{esquisser_ssb}
\title{Wrapper function for running esquisse in Jupyter at SSB}
\usage{
esquisser_ssb(data, ...)
}
\arguments{
\item{data}{Data frame to use for plotting}

\item{...}{Additional parameters for send to esquisser()}
}
\description{
Wrapper function for running esquisse in Jupyter at SSB
}
