% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\name{export_gcs}
\alias{export_gcs}
\title{Eksport av filer til GCS-bøtter med \code{rio::export}.}
\usage{
export_gcs(data, file, ...)
}
\arguments{
\item{data}{objekt som skal eksporteres.}

\item{file}{Full sti og filnavn på filen som skal opprettes. Eventuelt
\verb{gs://}-prefiks fjernes, og det øverste mappenivået tolkes som navn på
GCS-bøtten.}

\item{...}{Argumenter som sendes videre til \link[rio:export]{rio::export}, og dermed til
underliggende eksporterende funksjoner.}

\item{export_function}{En funksjon med parametere \code{input} og \code{output} som
eksporterer den ønskede filen.}
}
\value{
Resultatet av eksporten som returnert av \link[rio:export]{rio::export}.
}
\description{
Skriv data.frame og andre objekter til bøtter med  \link[rio:export]{rio::export}.
}
\details{
Parameterene \code{x}, \code{file}, og alle parametere i \code{...} blir sendt
videre til \link[rio:export]{rio::export} (default), eller en funksjon angitt .

Se \link[rio:export]{rio::export} for en oversikt over filformatene som kan eksporteres med
denne funksjonen.
}
\examples{
\dontrun{

export_gcs(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_xlsx_test.xlsx")

export_gcs(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_csv_test.csv")

export_gcs(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")

export_gcs(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_rds_test.rds")

export_gcs(data = list(foo = "bar", egg = "ham"),
           file = "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_json_test.json")

}

}
\seealso{
\code{\link[=import_gcs]{import_gcs()}} og \code{\link[rio:export]{rio::export()}}
}
