% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fellesr-package.R
\docType{package}
\name{fellesr-package}
\alias{fellesr}
\alias{fellesr-package}
\title{fellesr: General R functions for SSB}
\description{
This package is for gerenal R functions which may be useful for general use at SSB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statisticsnorway.github.io/fellesr/}
}

}
\author{
\strong{Maintainer}: Susie Jentoft \email{susie.jentoft@ssb.no}

Authors:
\itemize{
  \item Sindre Haugen \email{sindre.haugen@ssb.no}
}

}
\keyword{internal}
