% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{find_problem}
\alias{find_problem}
\title{Internal library function for a difficult package}
\usage{
find_problem(pkg, lib)
}
\arguments{
\item{pkg}{Character vector with the name of the package}

\item{lib}{Character vector with the path to the librarys}
}
\value{
Invisible `NULL`
}
\description{
Internal library function for a difficult package
}
\keyword{internal}
