% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{finn_versjon}
\alias{finn_versjon}
\title{Finn versjonsnummer fra et filnavn}
\usage{
finn_versjon(fil)
}
\arguments{
\item{fil}{En karakterstreng som representerer filnavnet hvor funksjonen skal søke etter et
versjonsnummer.}
}
\value{
Versjonsnummeret som et heltall (integer) dersom det finnes, ellers \code{FALSE} dersom filen
ikke følger versjonsstandard.
}
\description{
Funksjonen finner og returnerer versjonsnummeret fra et filnavn som følger en navnestandard
der versjonsnummeret er spesifisert etter \verb{_v}, etterfulgt av et eller flere siffer.
Dersom filen ikke er versjonert i henhold til denne standarden, returnerer funksjonen \code{FALSE}.
}
\details{
Funksjonen forventer at filnavnet inneholder et versjonsnummer etter mønsteret \verb{_v} fulgt av ett eller flere siffer.
For eksempel vil et filnavn som \code{rapport_v2.csv} returnere \code{2} som versjonsnummer.

Dersom filnavnet ikke inneholder et versjonsnummer i dette formatet, vil funksjonen returnere \code{FALSE}
og gi en advarsel om at filen ikke er versjonert i henhold til navnestandarden. Se
\href{https://manual.dapla.ssb.no/statistikkere/navnestandard.html}{navnestandarden} for mer informasjon.

Funksjonen gir også en advarsel dersom filnavnet inneholder stor "V" i stedet for liten "v", siden det forventes at
versjonsnummeret skal markeres med små bokstaver (\verb{_v}).
}
\examples{
# Finn versjonsnummer fra et gyldig filnavn
finn_versjon("rapport_v2.csv")  # Returnerer 2

# Filnavn uten versjonsnummer (advarsel vil gis)
finn_versjon("rapport.csv")  # Returnerer advarsel og FALSE

# Filnavn med stor "V" (advarsel vil gis)
finn_versjon("rapport_V2.csv")  # Returnerer advarsel og FALSE

}
