% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_package_template.R
\name{fix_file}
\alias{fix_file}
\title{Fix files by replacement of holders}
\usage{
fix_file(destination, file, find, replace)
}
\arguments{
\item{destination}{Where the file should be stored}

\item{file}{Target file}

\item{find}{Term to find}

\item{replace}{Replacement term}
}
\description{
Fix files by replacement of holders
}
