% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{fix_files}
\alias{fix_files}
\title{Fix several files}
\usage{
fix_files(
  path,
  package_name,
  prefixed_name,
  description,
  firstname,
  surname,
  email,
  type = "package"
)
}
\description{
Fix several files
}
\keyword{internal}
