% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{gcs.list.files}
\alias{gcs.list.files}
\title{Funksjon for å sjekke hvilke filer som finnes i en mappe i en Google Cloud Storage bucket}
\usage{
gcs.list.files(bucket)
}
\arguments{
\item{bucket}{Full sti til Google Cloud Storage bucket (med eventuelle undermapper).}
}
\description{
Funksjonen \code{gcs.list.files} kan brukes til å sjekke hvilke filer som finnes i en Google Cloud Storage bucket
}
\examples{
\dontrun{
gcs.list.files("ssb-prod-dapla-felles-data-delt/R_smoke_test")
}
}
