% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{gcs_delete_object}
\alias{gcs_delete_object}
\title{Funksjon for å slette fil fra en Google Cloud Storage bucket}
\usage{
gcs_delete_object(file)
}
\arguments{
\item{bucket}{Full sti til Google Cloud Storage bucket og filen som skal slettes.}
}
\description{
Funksjonen \code{gcs_delete_object} kan brukes til å slette filer fra Google Cloud Storage bucket.
}
\examples{
\dontrun{
gcs_delete_object("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")
}
}
