% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{gcs_list_objects}
\alias{gcs_list_objects}
\title{Funksjon for å sjekke hvilke filer som finnes i en Google Cloud Storage bucket}
\usage{
gcs_list_objects(bucket)
}
\arguments{
\item{bucket}{Full sti til Google Cloud Storage bucket.}
}
\description{
Funksjonen \code{list.files} kan brukes til å sjekke hvilke filer som finnes i en Google Cloud Storage bucket. I tillegg ser man størrelsen til filene og tidspnktet filen sist ble endret.
}
\examples{
\dontrun{
gcs_list_objects("ssb-prod-dapla-felles-data-delt/R_smoke_test")
}
}
