% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\name{gcs_validate_file_path}
\alias{gcs_validate_file_path}
\title{Kontroller formatet til en GCS-filsti}
\usage{
gcs_validate_file_path(path)
}
\arguments{
\item{path}{En karakter-vektor med lengde en.}
}
\value{
Dersom vektoren har riktig format, en navngitt karaktervektor med
lengde 4. Det første elementet i vektoren er lik \code{path}, og de andre
elementene har navn \code{bucket}, \code{file} og \code{extension}. Det første elementet
er den opprinnelige vektoren, det andre elementet er GCS-bøtten i filstien
uten ev. \verb{gs://}-prefiks, og det tredje elementet er resten av filstien,
inkludert mapper, filnavn og filendelse.

Dersom vektoren ikke har riktig format gir funksjonen en feilmelding.
}
\description{
Gir informasjon om bøtte, filsti og filendelse for gyldige GCS-filstier, og
feilmelding for ugyldige GCS-filstier.
}
\examples{

\dontrun{

gcs_validate_file_path()

}
}
