% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{get_files}
\alias{get_files}
\title{Copy files from standard template}
\usage{
get_files(path, project_type)
}
\arguments{
\item{path}{Path to new project}

\item{project_type}{Type of project. Choose between "project" or "package".}
}
\description{
Copy files from standard template
}
\keyword{internal}
