% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_format.R
\name{get_format}
\alias{get_format}
\title{Laster et SSB-format objekt fra JSON}
\usage{
get_format(filepath, is_range_format)
}
\arguments{
\item{filepath}{Sti til JSON-filen.}

\item{is_range_format}{Boolsk verdi som angir om formatet er intervallbasert.}
}
\value{
Et \code{ssb_format}-objekt.
}
\description{
Laster et SSB-format objekt fra JSON
}
\examples{
\dontrun{
format <- get_format("path/to/file.json", TRUE)
}
}
