% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{get_standard_files}
\alias{get_standard_files}
\title{Download standard files}
\usage{
get_standard_files(path)
}
\arguments{
\item{path}{Path to where the new project is being set up}
}
\description{
Download standard files
}
\keyword{internal}
