% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templater_funcs.R
\name{get_standard_files_offline}
\alias{get_standard_files_offline}
\title{Copy project files}
\usage{
get_standard_files_offline(path)
}
\arguments{
\item{path}{Path to where the new project is being set up}
}
\description{
Copy standard file to new project without accessing github standard files
}
\keyword{internal}
