% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eksport.R
\name{import_gcs}
\alias{import_gcs}
\title{Import av filer fra GCS-bøtter med \code{rio::import}.}
\usage{
import_gcs(file, ...)
}
\arguments{
\item{file}{Full sti og filnavn på filen som skal importeres. Eventuelt
\verb{gs://}-prefiks fjernes, og det øverste mappenivået tolkes som navn på
GCS-bøtten.}

\item{...}{Argumenter som sendes videre til \link[rio:import]{rio::import}, og dermed til
underliggende eksporterende funksjoner.}
}
\value{
Resultatet av å kjøre \link[rio:import]{rio::import} på den angitte filen.
}
\description{
Import av filer fra GCS-bøtter med \code{rio::import}.
}
\details{
Funksjonen laster ned den angitte filen på GCS til en midlertidig
fil i det lokale filsystemet, og benytter \link[rio:import]{rio::import} til å importere den.

Se \link[rio:import]{rio::import} for en oversikt over filformatene som kan importeres med
denne funksjonen.
}
\examples{

\dontrun{

import_gcs("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_xlsx_test.xlsx")
import_gcs("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_csv_test.csv")
import_gcs("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")
import_gcs("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_rds_test.rds")
import_gcs("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_json_test.json")


}

}
\seealso{
\code{\link[=gcs_export]{gcs_export()}} og \code{\link[rio:import]{rio::import()}}
}
