% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{install.packages}
\alias{install.packages}
\title{Install packages in SSB from CRAN
Wrapper function for utils::install.packages}
\usage{
install.packages(pkgs, lib, repos, ...)
}
\arguments{
\item{pkgs}{Name of the packages to install}

\item{lib}{Character vector giving library directory}

\item{repos}{Character vector giving CRAN repository to use}

\item{...}{Addition parameters (see https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/install.packages)}
}
\value{
Invisible \code{NULL}
}
\description{
Install packages in SSB from CRAN
Wrapper function for utils::install.packages
}
