% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{lag_versjonert_filsti}
\alias{lag_versjonert_filsti}
\title{Lag versjonert filsti}
\usage{
lag_versjonert_filsti(fil, versjon = c("siste", "ny", integer(1)))
}
\arguments{
\item{fil}{En karakterstreng som representerer filstien til filen som skal versjoneres.}

\item{versjon}{Angir hvordan versjonsnummeret skal håndteres. Kan være \code{"siste"} for å hente siste versjon,
\code{"ny"} for å lage en ny versjon, eller et heltall som spesifiserer en bestemt versjon. Standardverdien er \code{"siste"}.}
}
\value{
En karakterstreng som representerer filstien til den versjonerte filen. Funksjonen stopper med en feilmelding hvis
den ikke kan finne eller opprette en versjonert filsti i henhold til de angitte kriteriene.
}
\description{
Funksjonen genererer en filsti for en versjonert fil basert på et gitt filnavn. Den kan enten
finne siste versjon av en fil eller lage en ny versjon av filen.
}
\details{
Funksjonen tar et filnavn og sjekker om det er versjonert i henhold til et mønster der versjonsnummeret er angitt etter \verb{_v},
etterfulgt av ett eller flere siffer (f.eks. \verb{_v1}, \verb{_v2}). Den kan enten hente siste versjon, lage en ny versjon, eller returnere en
spesifikk versjon av filen hvis det er angitt.

Hvis filen ikke eksisterer i den angitte mappen, eller hvis den ikke inneholder en gyldig versjon, vil funksjonen stoppe med en feilmelding.

Når \code{versjon = "ny"}, vil funksjonen generere et nytt filnavn med versjonsnummer som er én høyere enn det høyeste eksisterende nummeret.
Hvis \code{versjon = "siste"}, returneres stien til den siste versjonen av filen. Hvis et spesifikt versjonsnummer er oppgitt som et heltall,
vil funksjonen returnere stien til denne versjonen hvis den eksisterer.
}
\examples{
\dontrun{
# Lag en ny versjon av filen "data.parquet" i mappen "prosjekt"
lag_versjonert_filsti("prosjekt/data_v1.parquet", versjon = "ny")

# Hent stien til siste versjon av filen "data.parquet"
lag_versjonert_filsti("prosjekt/data_v1.parquet", versjon = "siste")

# Hent stien til en spesifikk versjon (versjon 3) av filen "data.parquet"
lag_versjonert_filsti("prosjekt/data_v1.parquet", versjon = 3)

}
}
