% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{list.files}
\alias{list.files}
\title{Funksjon for å sjekke hvilke filer som finnes i en GCS bucket}
\usage{
list.files(bucket)
}
\arguments{
\item{bucket}{Full sti til GCS bucket.}
}
\description{
Funksjonen \code{list.files} kan brukes til å sjekke hvilke filer som finnes i en GCS bucket
}
\examples{
\dontrun{
list.files("ssb-prod-dapla-felles-data-delt/R_smoke_test")
}
}
