% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_feather}
\alias{read_feather}
\title{Funksjon for å laste inn .feather-fil fra Google Cloud Storage bucket}
\usage{
read_feather(file, ...)
}
\arguments{
\item{file}{Full sti og navn på filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_feather.html)}
}
\description{
Funksjonen \code{read_feather} kan brukes til å lese inn .feather-filer fra Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_feather("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.feather")
}
}
