% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eksport.R
\name{read_ssb}
\alias{read_ssb}
\alias{read_SSB}
\title{Import av filer på Dapla og på bakken}
\usage{
read_ssb(file, ...)

read_SSB(file, ...)
}
\arguments{
\item{file}{Full sti og filnavn på filen som skal importeres. Eventuelt
\verb{gs://}-prefiks fjernes, og det øverste mappenivået tolkes som navn på
GCS-bøtten.}

\item{...}{Argumenter som sendes videre til \link[rio:import]{rio::import}, og dermed til
underliggende eksporterende funksjoner.}
}
\description{
Import av filer på Dapla og på bakken
}
\examples{

\dontrun{

read_ssb("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_xlsx_test.xlsx")
read_ssb("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_csv_test.csv")
read_ssb("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")
read_ssb("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_rds_test.rds")
read_ssb("ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_json_test.json")


}

}
