% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_xml}
\alias{read_xml}
\title{Funksjon for aa laste inn .xml-fil fra Google Cloud Storage bucket}
\usage{
read_xml(file, ...)
}
\arguments{
\item{file}{Full sti og navn paa filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://rdrr.io/cran/googleCloudStorageR/man/gcs_get_object.html)}
}
\description{
Funksjonen \code{read_xml} kan brukes til aa lese inn .xml-filer fra Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_xml("ssb-prod-dapla-felles-data-delt/R_smoke_test/XXX.xml")
}
}
