% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{replace_file_path}
\alias{replace_file_path}
\title{Erstatt filsti i Datadoc-metadata}
\usage{
replace_file_path(filsti_datadoc, file_path)
}
\arguments{
\item{filsti_datadoc}{Tekststreng. Filsti til Datadoc JSON-filen som skal
oppdateres.}

\item{file_path}{Tekststreng. Ny filsti som skal lagres i
\code{datadoc$dataset$file_path}.}
}
\value{
Det oppdaterte Datadoc-objektet som en liste. Objektet returneres
usynlig.
}
\description{
Oppdaterer filstien som er lagret i datasettmetadataene i en Datadoc JSON-fil.
}
\details{
Funksjonen leser inn en eksisterende Datadoc JSON-fil, erstatter verdien i
\code{datadoc$dataset$file_path} og skriver de oppdaterte metadataene tilbake til
den samme JSON-filen.

Datadoc JSON-filen angitt i \code{filsti_datadoc} overskrives med den oppdaterte
versjonen. Den eksisterende verdien i \code{datadoc$dataset$file_path} erstattes
med verdien angitt i \code{file_path}.
}
\examples{
\dontrun{
replace_file_path(
  filsti_datadoc = "data/eksempel__DOC.json",
  file_path = "data/eksempel.parquet"
)
}

}
