% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{restore}
\alias{restore}
\title{Restore a SSB project
Wrapper function for renv::restore()}
\usage{
restore(repo, ...)
}
\arguments{
\item{repo}{Character vector pointing to which repository should be used}

\item{...}{}
}
\value{
Invisible `NULL`
}
\description{
Restore a SSB project
Wrapper function for renv::restore()
}
