% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runApp_ssb}
\alias{runApp_ssb}
\title{Wrapper function for running shiny apps in Jupyter at SSB}
\usage{
runApp_ssb(...)
}
\arguments{
\item{...}{Parameters to send to runApp()}
}
\value{
None
}
\description{
Wrapper function for running shiny apps in Jupyter at SSB
}
