% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{runExample_ssb}
\alias{runExample_ssb}
\title{Wrapper function for running exampleshiny apps in Jupyter at SSB}
\usage{
runExample_ssb(example, ...)
}
\arguments{
\item{...}{Parameters to send to runExample()}
}
\value{
None
}
\description{
Wrapper function for running exampleshiny apps in Jupyter at SSB
}
