% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ssb.R
\name{scale_colour_ssb}
\alias{scale_colour_ssb}
\alias{scale_colour_ssb,}
\alias{scale_color_ssb}
\title{Create continuous SSB colours}
\usage{
scale_colour_ssb(name = "main")
}
\arguments{
\item{name}{Name of the palette to use. Default is 'main'. Choose between:\cr
'main', greens', 'blues', 'primary', 'secondary', 'shade1', 'shade2','shade3', 'shade4', 'shade5'}
}
\value{
ggproto object for using in ggplot to change continuus colours
}
\description{
Create continuous SSB colours
}
