% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ssb.R
\name{scale_fill_ssb}
\alias{scale_fill_ssb}
\title{Create discete SSB colours}
\usage{
scale_fill_ssb(name = "main")
}
\arguments{
\item{name}{Name of the palette to use. Default is 'main'. Choose between:\cr
'main', greens', 'blues', 'primary', 'secondary', 'shade1', 'shade2','shade3', 'shade4', 'shade5'}
}
\value{
ggproto object for using in ggplot to change discrete colours
}
\description{
Create discete SSB colours
}
