% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{sjekk_endring}
\alias{sjekk_endring}
\title{Sjekk for endringer i kolonner, datatyper eller verdier mellom siste og nyeste versjon av en fil}
\usage{
sjekk_endring(filsti, versjon_1 = "uversjonert", versjon_2 = "siste")
}
\arguments{
\item{filsti}{En karakterstreng som representerer filstien til den nyeste versjonen av filen.
Funksjonen finner selv den siste versjonerte filen ved å bruke \code{\link{lag_versjonert_filsti}}.}
}
\value{
Funksjonen returnerer en logisk verdi:
\itemize{
\item \code{TRUE}: Hvis det har skjedd endringer i kolonnene, datatypene eller verdiene mellom den siste versjonerte filen og den nye filen.
\item \code{FALSE}: Hvis det ikke er noen endringer i noen av disse områdene.
}
}
\description{
Funksjonen sjekker om det har skjedd noen endringer i kolonner, datatyper eller verdier mellom den siste versjonerte filen og en ny fil.
Den kombinerer tre separate sjekker: for kolonner, datatyper og verdier.
}
\details{
Funksjonen kombinerer sjekker fra \code{\link{sjekk_endring_rader_kolonner}}, \code{\link{sjekk_endring_datatype}}, og \code{\link{sjekk_endring_verdier}}.
Den sjekker først om det har skjedd endringer i kolonnene, deretter i datatypene, og til slutt om det har skjedd endringer i verdiene mellom filene.
Hvis det oppdages en endring på noen av disse områdene, returnerer funksjonen \code{TRUE}. Hvis ingen endringer oppdages, returneres \code{FALSE}.
}
\examples{
\dontrun{
# Sjekk om det er noen endringer mellom siste versjon og ny fil
sjekk_endring("prosjekt/data.parquet")

}
}
