% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{sjekk_endring_datatype}
\alias{sjekk_endring_datatype}
\title{Sjekk for endringer i datatyper mellom siste og nyeste versjon av en fil}
\usage{
sjekk_endring_datatype(filsti, versjon_1 = "uversjonert", versjon_2 = "siste")
}
\arguments{
\item{filsti}{En karakterstreng som representerer filstien til den nyeste versjonen av filen.
Funksjonen finner selv den siste versjonerte filen ved å bruke \code{\link{lag_versjonert_filsti}}.}
}
\value{
Funksjonen returnerer en logisk verdi:
\itemize{
\item \code{TRUE}: Hvis det har skjedd endringer i datatypene til kolonnene mellom den siste versjonerte filen og den nye filen.
\item \code{FALSE}: Hvis det ikke er noen endringer i datatypene.
}
}
\description{
Funksjonen sammenligner datatypene til kolonnene i den siste versjonerte filen med datatypene i en ny fil
for å se om det har skjedd endringer i datatypene for kolonnene.
}
\details{
Funksjonen åpner den siste versjonerte filen i en gitt sti ved hjelp av \code{\link{lag_versjonert_filsti}} og sammenligner
kolonnenavn og datatyper med de i den nye filen som er angitt av \code{filsti}. Det kontrolleres om datatypene for
kolonnene har endret seg mellom den siste versjonen og den nye filen. Hvis det er en endring i datatypene, returnerer funksjonen \code{TRUE}.
Hvis ingen endringer oppdages, returneres \code{FALSE}.
}
\examples{
\dontrun{
# Sjekk om det er endringer i datatypene mellom siste versjon og ny fil
sjekk_endring_datatype("prosjekt/data.parquet")

}
}
