% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{sjekk_endring_verdier}
\alias{sjekk_endring_verdier}
\title{Sjekk for endringer i verdier mellom siste og nyeste versjon av en fil}
\usage{
sjekk_endring_verdier(filsti, versjon_1 = "uversjonert", versjon_2 = "siste")
}
\arguments{
\item{filsti}{En karakterstreng som representerer filstien til den nyeste versjonen av filen.
Funksjonen finner selv den siste versjonerte filen ved å bruke \code{\link{lag_versjonert_filsti}}.}
}
\value{
Funksjonen returnerer en logisk verdi:
\itemize{
\item \code{TRUE}: Hvis det har skjedd endringer i verdiene mellom den siste versjonerte filen og den nye filen.
\item \code{FALSE}: Hvis det ikke er noen endringer i verdiene.
}
}
\description{
Funksjonen sammenligner innholdet (verdiene) i den siste versjonerte filen med en ny fil for å se om det har skjedd endringer i dataene.
}
\details{
Funksjonen bruker \code{\link{dplyr::anti_join}} for å sammenligne verdiene i den siste versjonerte filen med den nye filen.
Den sjekker om det finnes rader som er forskjellige mellom de to datasettene. Hvis det finnes forskjeller, returnerer funksjonen \code{TRUE}.
Hvis ingen forskjeller oppdages, returneres \code{FALSE}.
}
\examples{
\dontrun{
# Sjekk om det er endringer i verdiene mellom siste versjon og ny fil
sjekk_endring_verdier("prosjekt/data.parquet")

}
}
