% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{snapshot}
\alias{snapshot}
\title{Snapshot a SSB project
Wrapper function for renv::snapshot}
\usage{
snapshot(ignore, ...)
}
\arguments{
\item{ignore}{List of packages to ignore}
}
\value{
None. Both lock and dependencies.R files will be updated. For more
details see <https://rstudio.github.io/renv/reference/snapshot.html>
}
\description{
Snapshot a SSB project
Wrapper function for renv::snapshot
}
