% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{ssb_library}
\alias{ssb_library}
\title{Library function for a difficult package
Some packages are not able to be installed by the user for various reasons. In this case the package may be able to be called from the common SSB library using this function.}
\usage{
ssb_library(package, ...)
}
\arguments{
\item{package}{Character vector with the name of the package}

\item{...}{}
}
\value{
Invisible \code{NULL}
}
\description{
Library function for a difficult package
Some packages are not able to be installed by the user for various reasons. In this case the package may be able to be called from the common SSB library using this function.
}
