% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ssb.R
\name{ssb_palettes}
\alias{ssb_palettes}
\title{Create SSB palette}
\usage{
ssb_palettes(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of the palette}

\item{n}{Number of colours}

\item{type}{Type of data. Can be "discrete" or "continuous"}
}
\value{
Vector of colours
}
\description{
Create SSB palette
}
