% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_project_template.R
\name{ssb_rproject}
\alias{ssb_rproject}
\title{SSB package template}
\usage{
ssb_rproject(path, description, firstname, surname, github)
}
\arguments{
\item{path}{Where the package should be created.}

\item{description}{A short description of what the package does.}

\item{firstname}{First name of the author}

\item{surname}{Surname of the Author}

\item{github}{Boolean for whether to create a github repository for the package. Default = TRUE}
}
\description{
Create an R package with standard SSB format
}
