% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{statbank_body}
\alias{statbank_body}
\title{statbank_body}
\usage{
statbank_body(data, tabell_id, ask = TRUE, boundary = 12345)
}
\arguments{
\item{...}{}
}
\description{
\code{statbank_body}
}
\examples{
\dontrun{
}
}
