% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{statbank_encrypt_request}
\alias{statbank_encrypt_request}
\title{statbank_encrypt_request}
\usage{
statbank_encrypt_request(laste_bruker)
}
\arguments{
\item{...}{}
}
\description{
\code{statbank_encrypt_request}
}
\examples{
\dontrun{
}
}
