% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{statbank_transfer}
\alias{statbank_transfer}
\title{Funksjon for å laste opp data fra DAPLA til Statistikkbanken}
\usage{
statbank_transfer(
  lastefil,
  lastefilsti = "",
  tabell_id,
  laste_bruker,
  publiseringsdato,
  initialer = gsub("@ssb.no", "", Sys.getenv("JUPYTERHUB_USER")),
  autooverskriv = 1,
  autogodkjenn = 2,
  boundary = 12345,
  ask = TRUE,
  validering = TRUE
)
}
\arguments{
\item{lastefil}{Objekt eller en liste med objekter som inneholder dataene som skal lastes opp. Det er også mulig å oppgi en karaktervektor med filnavn til .dat-filer lagret i en Google Cloud Storage bucket. Filstien oppgis under \code{lastefilsti}.}

\item{lastefilsti}{Karaktervektor med filsti til en Google Cloud Storage bucket hvor lastefilene (.dat) er plassert. Dersom man laster opp objekter direkte kan denne stå blank.}

\item{tabell_id}{Karaktervektor med tabell ID til tabellen som det skal lastes opp data til.}

\item{laste_bruker}{Karaktervektor med seksjonens lastebruker.}

\item{publiseringsdato}{Karaktervektor med publiseringsdato. Formatet skal være "ÅÅÅÅ-MM-DD".}

\item{initialer}{Karaktervektor med initialer til personen som laster opp dataene (og som mottar e-post med lastelogg). Standverdi hentes fra miljøvariabel i Jupyter så dersom det ikke er en annen person enn brukeren som kjører programmet som skal motta e-post kan denne stå blank.}

\item{autooverskriv}{Numerisk vektor. Standardverdi er satt til 1.}

\item{autogodkjenn}{Numerisk vektor. 0: manuell, 1: automatisk (umiddelbart), 2: JIT (just-in-time). Standardverdi er satt til 2.}

\item{boundary}{Numerisk vektor med tallverdien som skiller de ulike filene i opplastningen. Trenger ikke å endres.}

\item{ask}{Boolsk. Hvis \code{TRUE} blir man spurt om passord til lastebrukeren. Hvis \code{FALSE} må \code{username_encryptedpassword} først være laget med funksjonen \code{statbank_encrypt_request}.}

\item{validering}{}
}
\description{
Funksjonen \code{statbank_transfer} laster opp data fra DAPLA til Statistikkbanken. Det er mulig å laste opp objekter direkte eller ved å oppgi filsti til .dat-filer som er lagret i en Google Cloud Storage bucket.
}
\examples{
\dontrun{
transfer_log <- statbank_lasting(lastefil = roykalderkj1,
                                 tabell_id = "05307",
                                 laste_bruker = "LAST330",
                                 publiseringsdato = "2022-12-31")
transfer_log

transfer_log <- statbank_lasting(lastefil = "roykalderkj1.dat",
                                 lastefilsti = "ssb-prod-spesh-personell-data-kilde",
                                 tabell_id = "05307",
                                 laste_bruker = "LAST330",
                                 publiseringsdato = "2022-12-31")
transfer_log
}
}
