% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{statbank_uttaksbeskrivelse}
\alias{statbank_uttaksbeskrivelse}
\title{statbank_uttaksbeskrivelse}
\usage{
statbank_uttaksbeskrivelse(
  tabell_id,
  laste_bruker,
  ask = TRUE,
  boundary = 12345
)
}
\arguments{
\item{...}{}
}
\description{
\code{statbank_uttaksbeskrivelse}
}
\examples{
\dontrun{
}
}
