% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{statbank_validering}
\alias{statbank_validering}
\title{statbank_validering}
\usage{
statbank_validering(data, tabell_id, laste_bruker, ask = FALSE)
}
\arguments{
\item{...}{}
}
\description{
\code{statbank_validering}
}
\examples{
\dontrun{
}
}
