% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ssb.R
\name{theme_ssb}
\alias{theme_ssb}
\title{SSB theme for ggplot}
\usage{
theme_ssb()
}
\value{
theme object
}
\description{
SSB theme for ggplot
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x=cyl, fill=as.factor(cyl))) +
geom_bar( ) +
theme_ssb()
}
