% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{try_common_library}
\alias{try_common_library}
\title{Internal library function for a difficult package}
\usage{
try_common_library(pkg)
}
\arguments{
\item{pkg}{Character vector with the name of the package}
}
\value{
Invisible \code{NULL}
}
\description{
Internal library function for a difficult package
}
\keyword{internal}
