% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{vardef_get_klass_id}
\alias{vardef_get_klass_id}
\title{Hent KLASS-ID fra en variabeldefinisjon}
\usage{
vardef_get_klass_id(variable_definition)
}
\arguments{
\item{variable_definition}{En liste eller et listelignende objekt som
inneholder feltet \code{classification_uri}.}
}
\value{
En tekststreng med KLASS-ID-en. Dersom \code{classification_uri}
mangler, er tom eller ikke inneholder en numerisk ID på slutten av
adressen, returneres \code{NA_character_}.
}
\description{
Henter den numeriske KLASS-ID-en fra feltet \code{classification_uri} i en
variabeldefinisjon.
}
\details{
Funksjonen forventer at \code{classification_uri} avsluttes med en numerisk
KLASS-ID, eventuelt etterfulgt av en skråstrek.

Dersom \code{classification_uri} finnes, men ikke har forventet format, vises
en advarsel før \code{NA_character_} returneres.
}
\examples{
variable_definition <- list(
  classification_uri =
    "https://www.ssb.no/klass/klassifikasjoner/1"
)

vardef_get_klass_id(variable_definition)

variable_definition <- list(
  classification_uri =
    "https://www.ssb.no/klass/klassifikasjoner/1/"
)

vardef_get_klass_id(variable_definition)

variable_definition <- list(
  classification_uri = NULL
)

vardef_get_klass_id(variable_definition)

}
