% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{vardef_url}
\alias{vardef_url}
\title{Hent URL til tjenesten for variabeldefinisjoner}
\usage{
vardef_url(status = "ekstern")
}
\arguments{
\item{status}{En tekststreng som angir hvilken tjeneste URL-en skal
peke til. Gyldige verdier er \code{"ekstern"} og \code{"intern"}.}
}
\value{
En tekststreng med URL-en til den valgte tjenesten.
}
\description{
Returnerer endepunktet til SSBs tjeneste for variabeldefinisjoner.
}
\examples{
vardef_url()

vardef_url(status = "intern")

}
