% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{vars_with_value_labels}
\alias{vars_with_value_labels}
\title{Finn variabler med verdietiketter}
\usage{
vars_with_value_labels(data)
}
\arguments{
\item{data}{Et datasett som skal undersøkes for verdietiketter.}
}
\value{
En tekstvektor med navnene på variablene som har minst én
verdietikett. Dersom ingen variabler har verdietiketter, returneres
en tom tekstvektor.
}
\description{
Finner variabler i et datasett som har én eller flere
verdietiketter.
}
\details{
Verdietikettene for hver variabel hentes med
\code{\link[labelled:val_labels]{labelled::val_labels()}}. En variabel inkluderes i resultatet dersom
den har minst én registrert verdietikett.

Kolonneetiketter, som hentes med \code{\link[labelled:var_label]{labelled::var_label()}}, tas ikke med
i vurderingen.
}
\examples{
data <- data.frame(
  kjoenn = labelled::labelled(
    c(1, 2, 1),
    labels = c(
      Mann = 1,
      Kvinne = 2
    )
  ),
  alder = c(35, 42, 28)
)

vars_with_value_labels(data)

}
\seealso{
\code{\link[labelled:val_labels]{labelled::val_labels()}} for å hente verdietiketter og
\code{\link[=vars_without_labels]{vars_without_labels()}} for å finne variabler uten kolonneetikett.
}
