% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{vars_without_labels}
\alias{vars_without_labels}
\title{Finn variabler uten kolonneetikett}
\usage{
vars_without_labels(data)
}
\arguments{
\item{data}{Et datasett som skal undersøkes for kolonneetiketter.}
}
\value{
En tekstvektor med navnene på variablene som mangler
kolonneetikett. Dersom alle variablene har en kolonneetikett,
returneres en tom tekstvektor.
}
\description{
Finner variabler i et datasett som mangler en kolonneetikett.
}
\details{
Kolonneetikettene hentes med \code{\link[labelled:var_label]{labelled::var_label()}}.

En variabel regnes som å mangle kolonneetikett dersom etiketten:

\itemize{
\item er \code{NULL};
\item har lengde null;
\item bare inneholder manglende verdier; eller
\item bare inneholder tom tekst eller mellomrom.
}
}
\examples{
data <- data.frame(
  kjoenn = c(1, 2, 1),
  alder = c(35, 42, 28),
  inntekt = c(450000, 520000, 390000)
)

labelled::var_label(data$kjoenn) <- "Kjønn"
labelled::var_label(data$alder) <- "Alder i år"

vars_without_labels(data)

}
\seealso{
\code{\link[labelled:var_label]{labelled::var_label()}} for å hente eller angi kolonneetiketter.
}
