% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{versjoner_filer}
\alias{versjoner_filer}
\title{Versjonere filer}
\usage{
versjoner_filer(
  filstier,
  logg_fil = "versjonering_logg.json",
  arbeidsmappe,
  periode
)
}
\arguments{
\item{filstier}{En karaktervektor som inneholder navnene på filstier som skal versjoneres.
Funksjonen antar at filene er registrert som variable i det globale miljøet.}

\item{logg_fil}{Navnet på loggfilen der kjøringsinformasjon og versjonerte filer lagres. Standard er "versjonering_logg.json".}

\item{arbeidsmappe}{Stien til arbeidsmappen der loggfilen lagres.}
}
\value{
En liste som inneholder filstiene til de siste versjonene av de versjonerte filene.
}
\description{
Funksjonen oppdaterer filstier for flere filer ved å lage en ny versjon av filene dersom det er oppdaget endringer,
eller returnerer stien til den siste versjonerte filen hvis ingen endringer er oppdaget.
Funksjonen logger også resultatene av kjøringen ved hjelp av \code{\link{logg_kjoring}}, som lagrer informasjon om
versjonerte filer og kjøringsattributter som \code{kjoring_id} og \code{dato_tid} i en egen loggfil.
}
\details{
Funksjonen itererer over en liste med filstier og sjekker først om en versjonert fil allerede eksisterer. Hvis det ikke finnes
en versjonert fil opprettes versjon 1 (v1) automatisk. Dersom versjonerte filer allerede eksisterer sjekkes det om det har skjedd endringer
i kolonner, datatyper eller verdier for hver fil (mot den siste versjonerte filen) ved hjelp av \code{\link{sjekk_endring}}. Hvis det er oppdaget endringer,
vil funksjonen opprette en ny versjon av filen. Hvis ingen endringer oppdages, returneres stien til den siste versjonerte filen.
Etter hver kjøring, oppdateres loggfilen med informasjon om hvilke versjoner av filene som har blitt brukt i kjøringen samt metadata om kjøringen
(\code{kjoring_id} og \code{dato_tid}) ved hjelp av \code{\link{logg_kjoring}}.
}
\examples{
\dontrun{
# Lag versjonerte filstier for en liste med filnavn
filstier <- c("filsti1", "filsti2")
versjoner_filer(filstier)

}
}
