% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_dataset}
\alias{write_dataset}
\title{Funksjon for aa lagre "partitioned" .parquet-fil til Google Cloud Storage bucket}
\usage{
write_dataset(data, file, ...)
}
\arguments{
\item{data}{Filen som skal skrives.}

\item{file}{Full filsti og filnavn for hvor filen skal skrives.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/write_dataset.html)}
}
\description{
Funksjonen \code{write_dataset} kan brukes til aa skrive "partitioned" .parquet-filer til Google Cloud Storage bucket. "Partitioning" angis ut fra hvilke variabler datasettet er gruppert etter (gjoeres via \code{\link[dplyr:group_by]{dplyr::group_by()}}).
}
\examples{
\dontrun{
write_dataset(data, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_multifile_dataset_test")
}
}
