% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_parquet}
\alias{write_parquet}
\title{Funksjon for å lagre .parquet-fil til Google Cloud Storage bucket}
\usage{
write_parquet(data, file, ...)
}
\arguments{
\item{file}{Navn på filen som skal skrives.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_delim_arrow.html)}

\item{bucket}{Full sti til Google Cloud Storage bucket.}
}
\description{
Funksjonen \code{write_parquet} kan brukes til å skrive parquet-filer til Google Cloud Storage bucket.
}
\examples{
\dontrun{
}
}
