% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_rds}
\alias{write_rds}
\title{Funksjon for å lagre .rds-fil til Google Cloud Storage bucket}
\usage{
write_rds(data, file, ...)
}
\arguments{
\item{data}{Filen som skal skrives.}

\item{file}{Full filsti og filnavn for hvor filen skal skrives.}
}
\description{
Funksjonen \code{write_rds} kan brukes til å skrive .rds-filer til Google Cloud Storage bucket.
}
\examples{
\dontrun{
write_rds(data, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_rds_test.rds")
}
}
