% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_SSB}
\alias{write_SSB}
\title{Funksjon for å skrive filer til Google Cloud Storage bucket
#'
Funksjonen \code{write_SSB} kan brukes til å skrive filer til Google Cloud Storage.}
\usage{
write_SSB(data, file, partitioning = FALSE, ...)
}
\arguments{
\item{file}{Navn på filen som skal leses inn.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_parquet.html)}

\item{bucket}{Full sti til Google Cloud Storage bucket.}
}
\description{
Funksjon for å skrive filer til Google Cloud Storage bucket
#'
Funksjonen \code{write_SSB} kan brukes til å skrive filer til Google Cloud Storage.
}
\examples{
\dontrun{
}
}
