% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_funksjoner.R
\name{datadoc_path}
\alias{datadoc_path}
\title{Lag filsti til Datadoc-fil}
\usage{
datadoc_path(filsti)
}
\arguments{
\item{filsti}{En tekststreng eller tegnvektor med filstien til én eller
flere Parquet-filer.}
}
\value{
En tekststreng eller tegnvektor med filstien til den tilhørende
DataDoc-filen.
}
\description{
Oppretter filstien til en Datadoc-fil basert på filstien til en
Parquet-fil. Filendelsen \code{.parquet} erstattes med \verb{__DOC.json}.
}
\examples{
datadoc_path("/buckets/data/personell_v1.parquet")

datadoc_path(
  c(
    "/buckets/data/personell_v1.parquet",
    "/buckets/data/regnskap_v1.parquet"
  )
)

}
