% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{env_check}
\alias{env_check}
\title{Funksjon for aa sjekke hvilket miljoe man er i}
\usage{
env_check()
}
\value{
Karaktervektor: "DaplaProd", "DaplaTest", "BakkeProd", "BakkeTest" eller "Onyxia"
}
\description{
\code{env_check} er en hjelpefunksjon som sjekker hvilket miljoe man er i
}
\examples{
\dontrun{
env_check()
}
}
