% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\name{env_check}
\alias{env_check}
\title{Funksjon for å sjekke hvilket miljø man er i}
\usage{
env_check()
}
\description{
\code{env_check} er en hjelpefunksjon som sjekker hvilket miljø man er i (DaplaProd, DaplaTest, BakkeProd, BakkeTest eller Onyxia).
}
