% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{gcs_bucket}
\alias{gcs_bucket}
\title{Funksjon for å koble til GCS bucket}
\usage{
gcs_bucket(bucket)
}
\arguments{
\item{bucket}{Full sti til GCS bucket.}
}
\description{
\code{gcs_bucket} er en hjelpefunksjon som kobler til en GCS bucket ved hjelp av access_token og expiration som er lagret som miljøvariabler i Jupyter på DAPLA.
}
\examples{
\dontrun{
bucket <- gcs_bucket("ssb-prod-dapla-felles-data-delt")
bucket
}
}
