% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{gcs_global_bucket}
\alias{gcs_global_bucket}
\title{Funksjon for å koble til Google Cloud Storage bucket med googleCloudStorageR}
\usage{
gcs_global_bucket(bucket)
}
\arguments{
\item{bucket}{Full sti til Google Cloud Storage bucket.}
}
\description{
\code{gcs_global_bucket} er en hjelpefunksjon som kobler til en bucket på Google Cloud Storage med pakken \code{googleCloudStorageR}. Autentiseringen skjer via access_token og expiration som er lagret som miljøvariabler i Jupyter på DAPLA.
}
\examples{
\dontrun{
gcs_global_bucket("ssb-prod-dapla-felles-data-delt")
bucket
}
}
