% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_statbank.R
\encoding{UTF-8}
\name{initialer_funk}
\alias{initialer_funk}
\title{Funksjon for å hente initialene til pålogget bruker}
\usage{
initialer_funk(lastefil)
}
\value{
Karaktervektor med initialer (vanligvis tre siffer)
}
\description{
Funksjonen \code{initialer_funk} henter initialene til brukeren som er pålogget Jupyterlab/RStudio.
}
\examples{
\dontrun{
initialer_funk()
}
}
