% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{open_dataset}
\alias{open_dataset}
\title{Funksjon for å laste inn "multifile" datasett fra GCS bucket}
\usage{
open_dataset(bucket, ...)
}
\arguments{
\item{bucket}{Full sti til GCS bucket.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_parquet.html)}

\item{file}{Navn på filen som skal leses inn.}
}
\description{
Funksjonen \code{open_dataset} kan brukes til å lese deler av datasett (bl.a. .parquet-, .feather- og .csv-filer) fra GCS. Det lages en forbindelse til mappen der filen ligger og deretter kan man bruke argumenter fra \code{dplyr}, som \code{filter} og \code{select}, før man bruker \code{collect} til å lese inn dataene i R. \code{open_dataset} kan også brukes til å lese inn sf-objekter (lagret som .parquet-fil med pakken \code{sfarrow}).
}
\examples{
\dontrun{
}
}
