% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versjonering.R
\name{opprett_mappestruktur}
\alias{opprett_mappestruktur}
\title{Opprett mappestruktur}
\usage{
opprett_mappestruktur(
  arbeidsmappe,
  mapper = c("inndata", "klargjorte-data", "statistikk", "utdata"),
  periode = NULL
)
}
\arguments{
\item{arbeidsmappe}{En karakterstreng som spesifiserer hovedmappen der de nye undermappene skal opprettes.
Hvis denne stien slutter med en skråstrek (\code{/}), fjernes den automatisk. Dersom \code{arbeidsmappe} ikke
eksisterer, vil funksjonen automatisk opprette den.}

\item{mapper}{En karaktervektor som inneholder navnene på undermappene som skal opprettes. Standardverdien er
\code{c("inndata", "klargjorte-data", "statistikk", "utdata")}. Det er mulig å legge til flere mapper (f.eks. "oppdrag").}

\item{periode}{En valgfri karakterstreng som spesifiserer en undermappe som opprettes i hver av hovedmappene.
Hvis denne ikke er spesifisert (default er \code{NULL}), opprettes ingen slike undermapper.}
}
\value{
Ingen returverdi, men funksjonen oppretter mapper på filsystemet og lager variabler i det globale miljøet.
}
\description{
Funksjonen oppretter en standard mappestruktur basert på en filstien til en arbeidsmappe, og lager variabler
som peker på stien til hver undermappe i det globale miljøet. Det er også mulig å spesifisere en
valgfri parameter \code{periode} for å opprette en undermappe med det navnet i hver av hovedmappene.
}
\details{
Denne funksjonen går gjennom en liste av mapper, og for hver mappe:
\itemize{
\item Fjerner en eventuell avsluttende skråstrek (\code{/}) fra hovedmappen.
\item Oppretter \code{arbeidsmappe} dersom den ikke allerede finnes.
\item Oppretter undermappen inne i arbeidsmappen dersom den ikke allerede eksisterer.
\item Hvis \code{periode} er spesifisert, oppretter funksjonen også en undermappe med navnet spesifisert av \code{periode}
i hver hovedmappe.
\item Erstatt bindestreker i undermappenavn med understrek for å lage gyldige variabelnavn.
\item Tilordner hver undermappe-sti som en variabel i det globale miljøet (bruker \code{assign}).
}

Variabelnavnene som opprettes har formatet \code{mappe_navn_mappe}, der eventuelle bindestreker i mappe-navn
erstattes med understrek.
}
\examples{
# Opprett standard mappestruktur i en mappe "prosjekt"
opprett_mappestruktur("prosjekt")

# Opprett en egendefinert mappestruktur
opprett_mappestruktur("analyse", mapper = c("data", "output", "figurer"))

# Opprett mappestruktur med en spesifikk periode (f.eks. "2024" eller "2024-Q1")
opprett_mappestruktur("prosjekt", periode = "2024")

}
