% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_csv}
\alias{read_csv}
\title{Funksjon for aa laste inn .csv-fil fra Google Cloud Storage bucket}
\usage{
read_csv(file, ...)
}
\arguments{
\item{file}{Full sti og navn paa filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_delim_arrow.html)}
}
\description{
Funksjonen \code{read_csv} kan brukes til aa lese inn .csv-filer Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_csv("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.csv")
}
}
