% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_csv}
\alias{read_csv}
\title{Funksjon for å laste inn .csv-fil fra GCS bucket}
\usage{
read_csv(bucket, file, ...)
}
\arguments{
\item{bucket}{Full sti til GCS bucket.}

\item{file}{Navn på filen som skal leses inn.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_delim_arrow.html)}
}
\description{
Funksjonen \code{read_csv} kan brukes til å lese inn JSON-filer GCS.
}
\examples{
\dontrun{
data <- read_csv(bucket = "ssb-prod-spesh-personell-data-kilde", file = "example_1.csv")
}
}
