% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_feather}
\alias{read_feather}
\title{Funksjon for å laste inn .feather-fil fra GCS bucket
#'
Funksjonen \code{read_feather} kan brukes til å lese inn .feather-filer fra GCS.}
\usage{
read_feather(file, ...)
}
\arguments{
\item{file}{Navn på filen som skal leses inn.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_feather.html)}

\item{bucket}{Full sti til GCS bucket.}
}
\description{
Funksjon for å laste inn .feather-fil fra GCS bucket
#'
Funksjonen \code{read_feather} kan brukes til å lese inn .feather-filer fra GCS.
}
\examples{
\dontrun{
}
}
