% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_json}
\alias{read_json}
\title{Funksjon for å laste inn .json-fil fra Google Cloud Storage}
\usage{
read_json(file, ...)
}
\arguments{
\item{file}{Full sti og navn på filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_json_arrow.html)}
}
\description{
Funksjonen \code{read_json} kan brukes til å lese inn .json-filer Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_json("ssb-prod-spesh-personell-data-kilde/example_1.json")
}
}
