% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_parquet}
\alias{read_parquet}
\title{Funksjon for å laste inn .parquet-fil fra GCS bucket
#'
Funksjonen \code{read_parquet} kan brukes til å lese inn parquet-filer GCS.}
\usage{
read_parquet(bucket, file, ...)
}
\arguments{
\item{bucket}{Full sti til GCS bucket.}

\item{file}{Navn på filen som skal leses inn.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_parquet.html)}
}
\description{
Funksjon for å laste inn .parquet-fil fra GCS bucket
#'
Funksjonen \code{read_parquet} kan brukes til å lese inn parquet-filer GCS.
}
\examples{
\dontrun{
data <- read_parquet(bucket = "ssb-prod-dapla-felles-data-delt/R_smoke_test", file = "1987")

data <- read_parquet(bucket = "ssb-prod-dapla-felles-data-delt/R_smoke_test", file = "1987", col_select = c("Year", "Month"))
}
}
