% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_parquet_sf}
\alias{read_parquet_sf}
\title{Funksjon for å laste inn .parquet-fil (i sf-format) fra Google Cloud Storage bucket}
\usage{
read_parquet_sf(file, ...)
}
\arguments{
\item{file}{Full sti og navn på filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/read_parquet.html)}
}
\description{
Funksjonen \code{read_parquet_sf} kan brukes til å lese inn .parquet-filer (i sf-format) fra Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_parquet_sf("ssb-prod-dapla-felles-data-delt/GIS/testdata/veger_oslo.parquet")
}
}
