% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_rds}
\alias{read_rds}
\title{Funksjon for å laste inn .rds-fil fra Google Cloud Storage bucket}
\usage{
read_rds(file, ...)
}
\arguments{
\item{file}{Full sti og navn på filen som skal leses inn fra Google Cloud Storage bucket.}

\item{...}{Flere parametere (se: https://rdrr.io/cran/googleCloudStorageR/man/gcs_get_object.html)}
}
\description{
Funksjonen \code{read_rds} kan brukes til å lese inn .rds-filer fra Google Cloud Storage.
}
\examples{
\dontrun{
data <- read_rds("ssb-prod-spesh-personell-data-kilde/nodes.RDS")
}
}
