% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{read_SSB}
\alias{read_SSB}
\title{Funksjon for å laste inn filer fra Google Cloud Storage bucket}
\usage{
read_SSB(file, sf = FALSE, ...)
}
\arguments{
\item{file}{Full sti og navn på filen som skal leses inn fra Google Cloud Storage bucket.}

\item{sf}{Boolsk. Standardverdi er FALSE. Sett \code{sf = TRUE} dersom .parquet-filen er et sf-objekt.}

\item{...}{Flere parametere (se dokumentasjonen til: \code{\link[=read_parquet]{read_parquet()}}/\code{\link[=open_dataset]{open_dataset()}}/\code{\link[=read_feather]{read_feather()}}/\code{\link[=read_csv]{read_csv()}}/\code{\link[=read_rds]{read_rds()}})}
}
\description{
Funksjonen \code{read_SSB} kan brukes til å lese inn filer fra Google Cloud Storage. Funksjonen støtter .parquet- (inkludert sf-objekter), .feather-, .rds- og .csv- og .json-filer.
}
\examples{
\dontrun{
read_SSB_parquet <- read_SSB("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.parquet")

read_SSB_MFD <- read_SSB("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987_1996_dataset")

read_SSB_feather <- read_SSB("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.feather")

read_SSB_csv <- read_SSB("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.csv")

read_SSB_rds <- read_SSB("ssb-prod-dapla-felles-data-delt/R_smoke_test/1987.rds")

read_SSB_json <- read_SSB("ssb-prod-spesh-personell-data-kilde/example_1.json")

}
}
