% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_format.R
\name{ssb_format}
\alias{ssb_format}
\title{SSB-format}
\description{
Klassen \code{ssb_format} gir funksjonalitet for å mappe kategoriske og intervallbaserte verdier
i henhold til en forhåndsdefinert kodeliste.
}
\examples{

## ------------------------------------------------
## Method `ssb_format$map_range`
## ------------------------------------------------

\dontrun{
format <- ssb_format$new(dict = list("10-20" = "A", "21-30" = "B", "other" = "C"), is_range_format = TRUE)
format$map_range(c(15, 25, 35))
}

## ------------------------------------------------
## Method `ssb_format$map_cat`
## ------------------------------------------------

\dontrun{
format <- ssb_format$new(dict = list("A" = "Alpha", "B" = "Beta"), is_range_format = FALSE)
format$map_cat(c("A", "B"))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dict}}{En navngitt liste som spesifiserer mappingen.}

\item{\code{is_range_format}}{Boolsk verdi som angir om formatet er intervallbasert.}

\item{\code{na_value}}{Verdi som representerer manglende data utledet fra \code{dict}.}

\item{\code{breaks}}{En vektor av grenseverdier for intervallbasert mapping utledet fra \code{dict}.}

\item{\code{labels}}{En vektor med labeler for intervallene utledet fra \code{dict}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ssb_format-new}{\code{ssb_format$new()}}
\item \href{#method-ssb_format-map_range}{\code{ssb_format$map_range()}}
\item \href{#method-ssb_format-map_cat}{\code{ssb_format$map_cat()}}
\item \href{#method-ssb_format-clone}{\code{ssb_format$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssb_format-new"></a>}}
\if{latex}{\out{\hypertarget{method-ssb_format-new}{}}}
\subsection{Method \code{new()}}{
Initialiserer SSB Format-klassen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ssb_format$new(dict, is_range_format)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dict}}{En "named list" (tilsvarende dictionary i python) som spesifiserer mappingen.}

\item{\code{is_range_format}}{Boolsk verdi som angir om formatet er intervallbasert.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssb_format-map_range"></a>}}
\if{latex}{\out{\hypertarget{method-ssb_format-map_range}{}}}
\subsection{Method \code{map_range()}}{
Mapper numeriske verdier til kategoriske
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ssb_format$map_range(vec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{En vektor med numeriske verdier som skal mappes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
En vektor med kategoriske verdier som er mappet.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
format <- ssb_format$new(dict = list("10-20" = "A", "21-30" = "B", "other" = "C"), is_range_format = TRUE)
format$map_range(c(15, 25, 35))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssb_format-map_cat"></a>}}
\if{latex}{\out{\hypertarget{method-ssb_format-map_cat}{}}}
\subsection{Method \code{map_cat()}}{
Mapper kategorier til kategorier
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ssb_format$map_cat(vec)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vec}}{En vektor med kategoriske verdier som skal mappes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
En vektor med de mappede verdiene.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
format <- ssb_format$new(dict = list("A" = "Alpha", "B" = "Beta"), is_range_format = FALSE)
format$map_cat(c("A", "B"))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ssb_format-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ssb_format-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ssb_format$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
