% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_funcs.R
\name{ssb_library}
\alias{ssb_library}
\title{library function for a difficult package
Some packages are not abel to be installed by the user. In this case the package is intalled from the felles library}
\usage{
ssb_library(package, ...)
}
\arguments{
\item{package}{Character vector with the name of the package}

\item{...}{}
}
\value{
Invisible `NULL`
}
\description{
library function for a difficult package
Some packages are not abel to be installed by the user. In this case the package is intalled from the felles library
}
