% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statbankr_functions.R
\name{statbank_lasting}
\alias{statbank_lasting}
\title{Uploading to Statbank
Function for uploading a file in .dat format to the Norwegian Statistics Bank}
\usage{
statbank_lasting(
  lastefil,
  tabell_id = "",
  laste_dbase = "",
  laste_bruker,
  lastefilsti = "",
  hovedtabell = "",
  publiseringsdato,
  mailto = "",
  autooverskriv = 1,
  autogodkjenn = 1,
  timeout = 10,
  validering = TRUE,
  ask = TRUE
)
}
\arguments{
\item{lastefil}{Name of the .dat file for loading (case sensitive)}

\item{laste_dbase}{Database for loading up file. Choose between "TEST", "QA" or "PROD"}

\item{laste_bruker}{Oracle use ID for loading to StatBank}

\item{lastefilsti}{Path to where the .dat file is located}

\item{hovedtabell}{Name of the tabell for publication}

\item{publiseringsdato}{Date for publication. This should be in the format 'yyyy-mm-dd'. For example "2021-02-28".}

\item{mailto}{Initials (three letters) for the person to send confirmation email to. SSB initials. Default is user initials.}

\item{autooverskriv}{If copies should be overwritten. 0 = no, 1 = yes (default)}

\item{autogodkjenn}{How data should be approved. 0 = Manual approval, 1 = Automatic (default), 2 = Just-in-time}

\item{timeout}{Length of time to wait for script to run before stopping. Default 10 (seconds)}
}
\description{
Uploading to Statbank
Function for uploading a file in .dat format to the Norwegian Statistics Bank
}
