% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_feather}
\alias{write_feather}
\title{Funksjon for å lagre .feather-fil til Google Cloud Storage bucket}
\usage{
write_feather(data, file, ...)
}
\arguments{
\item{data}{Filen som skal skrives.}

\item{file}{Full filsti og filnavn for hvor filen skal skrives.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/write_feather.html)}
}
\description{
Funksjonen \code{write_feather} kan brukes til å skrive .feather-filer til Google Cloud Storage bucket.
}
\examples{
\dontrun{
write_feather(data, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.feather")
}
}
