% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_parquet}
\alias{write_parquet}
\title{Funksjon for aa lagre .parquet-fil til Google Cloud Storage bucket}
\usage{
write_parquet(data, file, ...)
}
\arguments{
\item{data}{Filen som skal skrives.}

\item{file}{Full filsti og filnavn for hvor filen skal skrives.}

\item{...}{Flere parametere (se: https://arrow.apache.org/docs/r/reference/write_parquet.html)}
}
\description{
Funksjonen \code{write_parquet} kan brukes til aa skrive .parquet-filer til Google Cloud Storage bucket.
}
\examples{
\dontrun{
write_parquet(data, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")
}
}
