% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPLA_funcs.R
\encoding{UTF-8}
\name{write_sf_parquet}
\alias{write_sf_parquet}
\title{Funksjon for å lagre sf-objekt som .parquet-fil til Google Cloud Storage bucket}
\usage{
write_sf_parquet(data, file)
}
\arguments{
\item{data}{Filen som skal skrives.}

\item{file}{Full filsti og filnavn for hvor filen skal skrives.}
}
\description{
Funksjonen \code{write_sf_parquet} kan brukes til å skrive sf-objekter som .parquet-filer til Google Cloud Storage bucket.
}
\examples{
\dontrun{
write_sf_parquet(data, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_sf.parquet")
}
}
