% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eksport.R
\name{write_ssb}
\alias{write_ssb}
\alias{write_SSB}
\title{Eksport av filer på Dapla og på bakken}
\usage{
write_ssb(data, file, ...)

write_SSB(data, file, ...)
}
\arguments{
\item{data}{objekt som skal eksporteres.}

\item{file}{Full sti og filnavn på filen som skal opprettes. Eventuelt
\verb{gs://}-prefiks fjernes, og det øverste mappenivået tolkes som navn på
GCS-bøtten.}

\item{...}{Argumenter som sendes videre til \link[rio:export]{rio::export}, og dermed til
underliggende eksporterende funksjoner.}
}
\value{
På bakken: resultatet av eksporten som returnert av \link[rio:export]{rio::export}.

På Dapla: resultatet av eksporten som returnert av \link[googleCloudStorageR:gcs_upload]{googleCloudStorageR::gcs_upload}
}
\description{
Eksport av filer på Dapla og på bakken
}
\details{
Se \link[rio:export]{rio::export} for tilgjengelige filformater. Funksjonen støtter de samme paramete
}
\examples{
\dontrun{

write_ssb(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_xlsx_test.xlsx")

write_ssb(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_csv_test.csv")

write_ssb(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_parquet_test.parquet")

write_ssb(mtcars, "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_rds_test.rds")

write_ssb(data = list(foo = "bar", egg = "ham"),
          file = "ssb-prod-dapla-felles-data-delt/R_smoke_test/write_SSB_json_test.json")

}

}
